Separate services into module files, add descriptive comment to tag parsing and fix compiler warnings
